+++ activerecord/test/cases/tasks/database_tasks_test.rb
@@ -251,43 +251,160 @@ module ActiveRecord
     end
   end
 
+  module SchemaFormatTestHelpers
+    def setup
+      @configurations = { "development" => { "primary" => { "database" => "dev-db" } } }
+    end
+
+    private
+      def with_stubbed_db_dir
+        Dir.mktmpdir do |dir|
+          ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, dir) do
+            yield dir
+          end
+        end
+      end
+
+      def with_stubbed_configurations
+        old_configurations = ActiveRecord::Base.configurations
+        ActiveRecord::Base.configurations = @configurations
+      ensure
+        ActiveRecord::Base.configurations = old_configurations
+        ActiveRecord::Base.clear_cache!
+      end
+  end
+
   class DatabaseTasksDumpSchemaTest < ActiveRecord::TestCase
+    include SchemaFormatTestHelpers
+
     def test_ensure_db_dir
-      Dir.mktmpdir do |dir|
-        ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, dir) do
-          updated_hash = ActiveRecord::Base.configurations.configs_for(env_name: "arunit", name: "primary").configuration_hash.merge(schema_dump: "fake_db_config_schema.rb")
-          db_config = ActiveRecord::DatabaseConfigurations::HashConfig.new("arunit", "primary", updated_hash)
+      with_stubbed_db_dir do |dir|
+        @configurations["development"]["primary"]["schema_dump"] = "fake_db_config_schema.rb"
+        with_stubbed_configurations do
           path = "#{dir}/fake_db_config_schema.rb"
 
-          FileUtils.rm_rf(dir)
+          FileUtils.rm_rf(path)
           assert_not File.file?(path)
 
+          db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
           ActiveRecord::Tasks::DatabaseTasks.dump_schema(db_config)
 
           assert File.file?(path)
         end
       end
-    ensure
-      ActiveRecord::Base.clear_cache!
     end
 
     def test_db_dir_ignored_if_included_in_schema_dump
-      Dir.mktmpdir do |dir|
-        ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, dir) do
-          updated_hash = ActiveRecord::Base.configurations.configs_for(env_name: "arunit", name: "primary").configuration_hash.merge(schema_dump: "#{dir}/fake_db_config_schema.rb")
-          db_config = ActiveRecord::DatabaseConfigurations::HashConfig.new("arunit", "primary", updated_hash)
+      with_stubbed_db_dir do |dir|
+        @configurations["development"]["primary"]["schema_dump"] = "#{dir}/fake_db_config_schema.rb"
+        with_stubbed_configurations do
           path = "#{dir}/fake_db_config_schema.rb"
 
-          FileUtils.rm_rf(dir)
+          FileUtils.rm_rf(path)
           assert_not File.file?(path)
 
+          db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
           ActiveRecord::Tasks::DatabaseTasks.dump_schema(db_config)
 
           assert File.file?(path)
         end
       end
-    ensure
-      ActiveRecord::Base.clear_cache!
+    end
+
+    def test_default_format_schema_dump
+      with_stubbed_db_dir do
+        with_stubbed_configurations do
+          assert_called(ActiveRecord::SchemaDumper, :dump) do
+            db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
+            ActiveRecord::Tasks::DatabaseTasks.dump_schema(db_config)
+          end
+        end
+      end
+    end
+
+    def test_config_override_ruby_format_schema_dump
+      @configurations["development"]["primary"]["schema_format"] = "ruby"
+
+      with_stubbed_db_dir do
+        with_stubbed_configurations do
+          assert_called(ActiveRecord::SchemaDumper, :dump) do
+            db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
+            ActiveRecord::Tasks::DatabaseTasks.dump_schema(db_config)
+          end
+        end
+      end
+    end
+
+    def test_config_override_sql_format_schema_dump
+      @configurations["development"]["primary"]["schema_format"] = "sql"
+
+      with_stubbed_db_dir do
+        with_stubbed_configurations do
+          assert_called(ActiveRecord::Tasks::DatabaseTasks, :structure_dump) do
+            db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
+            ActiveRecord::Tasks::DatabaseTasks.dump_schema(db_config)
+          end
+        end
+      end
+    end
+
+    def test_schema_dump_with_schema_format_param_is_deprecated
+      with_stubbed_db_dir do
+        with_stubbed_configurations do
+          db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
+          assert_deprecated { ActiveRecord::Tasks::DatabaseTasks.dump_schema(db_config, :ruby) }
+        end
+      end
+    end
+  end
+
+  class DatabaseTasksLoadSchemaTest < ActiveRecord::TestCase
+    include SchemaFormatTestHelpers
+
+    def test_default_format_schema_load
+      with_stubbed_db_dir do
+        with_stubbed_configurations do
+          assert_called(ActiveRecord::Tasks::DatabaseTasks, :load) do
+            db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
+            ActiveRecord::Tasks::DatabaseTasks.load_schema(db_config)
+          end
+        end
+      end
+    end
+
+    def test_schema_load_with_schema_format_param_is_deprecated
+      with_stubbed_db_dir do
+        with_stubbed_configurations do
+          db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
+          assert_deprecated { ActiveRecord::Tasks::DatabaseTasks.load_schema(db_config, :ruby) }
+        end
+      end
+    end
+
+    def test_config_override_ruby_format_schema_load
+      @configurations["development"]["primary"]["schema_format"] = "ruby"
+
+      with_stubbed_db_dir do
+        with_stubbed_configurations do
+          assert_called(ActiveRecord::Tasks::DatabaseTasks, :load) do
+            db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
+            ActiveRecord::Tasks::DatabaseTasks.load_schema(db_config)
+          end
+        end
+      end
+    end
+
+    def test_config_override_sql_format_schema_load
+      @configurations["development"]["primary"]["schema_format"] = "sql"
+
+      with_stubbed_db_dir do
+        with_stubbed_configurations do
+          assert_called(ActiveRecord::Tasks::DatabaseTasks, :structure_load) do
+            db_config = ActiveRecord::Base.configurations.configs_for(env_name: "development", name: "primary")
+            ActiveRecord::Tasks::DatabaseTasks.load_schema(db_config)
+          end
+        end
+      end
     end
   end
 
@@ -1578,10 +1695,23 @@ module ActiveRecord
     end
 
     { ruby: "schema.rb", sql: "structure.sql" }.each_pair do |fmt, filename|
-      define_method("test_check_dump_filename_for_#{fmt}_format") do
+      define_method("test_check_schema_dump_path_for_#{fmt}_format") do
+        ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "/tmp") do
+          configurations = {
+            "development" => { "primary" => { "database" => "dev-db", "schema_format" => fmt } }
+          }
+          with_stubbed_configurations(configurations) do
+            assert_equal "/tmp/#{filename}", ActiveRecord::Tasks::DatabaseTasks.schema_dump_path(config_for("development", "primary"))
+          end
+        end
+      end
+
+      define_method("test_schema_dump_path_with_explicit_#{fmt}_format_is_deprecated") do
         ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "/tmp") do
           with_stubbed_configurations do
-            assert_equal "/tmp/#{filename}", ActiveRecord::Tasks::DatabaseTasks.schema_dump_path(config_for("development", "primary"), fmt)
+            assert_deprecated do
+              assert_equal "/tmp/#{filename}", ActiveRecord::Tasks::DatabaseTasks.schema_dump_path(config_for("development", "primary"), fmt)
+            end
           end
         end
       end
@@ -1628,10 +1758,10 @@ module ActiveRecord
       define_method("test_check_dump_filename_for_#{fmt}_format_with_non_primary_databases") do
         ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "/tmp") do
           configurations = {
-            "development" => { "primary" => { "database" => "dev-db" }, "secondary" => { "database" => "secondary-dev-db" } },
+            "development" => { "primary" => { "database" => "dev-db" }, "secondary" => { "database" => "secondary-dev-db", "schema_format" => fmt } }
           }
           with_stubbed_configurations(configurations) do
-            assert_equal "/tmp/secondary_#{filename}", ActiveRecord::Tasks::DatabaseTasks.schema_dump_path(config_for("development", "secondary"), fmt)
+            assert_equal "/tmp/secondary_#{filename}", ActiveRecord::Tasks::DatabaseTasks.schema_dump_path(config_for("development", "secondary"))
           end
         end
       end